lab13/pyproject.toml
[project]
name = "MLOps_course_AGH_lab_13"
version = "1.0.0"
description = ""
readme = "README.md"

requires-python = "==3.11.*"

dependencies = [
    "fastmcp>=2.14.3",
    "google-genai>=1.61.0",
    "guardrails-ai==0.7.2",
    "litellm>=1.80.0",
    "matplotlib>=3.10.8",
    "mcp-use>=1.5.2",
    "openai<2", # Guardrails AI limitation
    "polars>=1.37.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "transformers[torch]>=4.51.0",
]

[dependency-groups]
dev = [
    "ruff"
]

cuda = [
    "bitsandbytes>=0.49.1",
    "vllm>=0.13.0"
]
